+++ node_pool_taint/main.tf
@@ -13,16 +13,16 @@
 ## a -/+ operation for the nodepool.
 # https://github.com/terraform-providers/terraform-provider-google/issues/1054
 resource "random_id" "entropy" {
-  keepers {
-    machine_type       = "${var.machine_type}"
-    name               = "${var.name}"
-    region             = "${var.region}"
-    disk_size          = "${var.disk_size_in_gb}"
-    tags               = "${join(",", sort(var.node_tags))}"
-    disk_type          = "${var.disk_type}"
-    labels             = "${jsonencode(var.node_labels)}"
-    taint              = "${jsonencode(var.taint)}"
-    initial_node_count = "${var.initial_node_count}"
+  keepers = {
+    machine_type       = var.machine_type
+    name               = var.name
+    region             = var.region
+    disk_size          = var.disk_size_in_gb
+    tags               = join(",", sort(var.node_tags))
+    disk_type          = var.disk_type
+    labels             = jsonencode(var.node_labels)
+    taint              = jsonencode(var.taint)
+    initial_node_count = var.initial_node_count
   }
 
   byte_length = 2
@@ -38,31 +38,44 @@ locals {
 }
 
 resource "google_container_node_pool" "node_pool" {
-  provider           = "google-beta"
+  provider           = google-beta
   name               = "${var.name}-${random_id.entropy.hex}"
-  cluster            = "${var.gke_cluster_name}"
-  location           = "${var.region}"
-  version            = "${var.kubernetes_version}"
-  initial_node_count = "${var.initial_node_count}"
+  cluster            = var.gke_cluster_name
+  location           = var.region
+  version            = var.kubernetes_version
+  initial_node_count = var.initial_node_count
 
   autoscaling {
-    min_node_count = "${var.min_node_count}"
-    max_node_count = "${var.max_node_count}"
+    min_node_count = var.min_node_count
+    max_node_count = var.max_node_count
   }
 
   node_config {
     image_type   = "COS"
-    disk_size_gb = "${var.disk_size_in_gb}"
-    machine_type = "${var.machine_type}"
-    labels       = "${var.node_labels}"
-    taint        = ["${var.taint}"]
-    disk_type    = "${var.disk_type}"
-    tags         = ["${var.node_tags}"]
+    disk_size_gb = var.disk_size_in_gb
+    machine_type = var.machine_type
+    labels       = var.node_labels
+    dynamic "taint" {
+      for_each = [var.taint]
+      content {
+        # TF-UPGRADE-TODO: The automatic upgrade tool can't predict
+        # which keys might be set in maps assigned here, so it has
+        # produced a comprehensive set here. Consider simplifying
+        # this after confirming which keys can be set in practice.
 
-    oauth_scopes = "${concat(local.base_oauth_scope, var.additional_oauth_scopes)}"
+        effect = taint.value.effect
+        key    = taint.value.key
+        value  = taint.value.value
+      }
+    }
+    disk_type = var.disk_type
+    tags      = var.node_tags
+
+    oauth_scopes = concat(local.base_oauth_scope, var.additional_oauth_scopes)
   }
 
   lifecycle {
     create_before_destroy = true
   }
 }
+