+++ vpc-native/main.tf
@@ -24,7 +24,6 @@ resource "google_container_cluster" "cluster" {
     }
   }
 
-  # Private cluster settings, default false
   private_cluster_config {
     enable_private_endpoint = var.enable_private_endpoint
     enable_private_nodes    = var.enable_private_nodes