+++ vpc-native/versions.tf
@@ -1,0 +1,7 @@
+
+terraform {
+  required_version = ">= 0.12"
+  required_providers {
+    google = ">=2.5.0"
+  }
+}